pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarma"
version = "0.1.0"
description = "Learning agent swarms that get smarter every cycle"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Rabin", email = "rabin@theglitch.ai" },
]
keywords = ["agents", "swarm", "llm", "orchestration", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "apscheduler>=3.10",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dashboard = ["fastapi>=0.110", "uvicorn>=0.29"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
swarma = "swarma.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["swarma"]

[tool.ruff]
target-version = "py311"
line-length = 100
